Refactor New Project dialog into separate module. Implement standard event handler initialization.
