+++ src/modules/modals.js
@@ -29,236 +29,14 @@ import * as Cookies from 'js-cookie';
 import {LOCAL_PROJECT_STORE_NAME} from './constants.js';
 import {TEMP_PROJECT_STORE_NAME} from './constants.js';
 
-import {createNewProject, isProjectChanged, insertProject} from './editor.js';
+import {isProjectChanged, insertProject} from './editor.js';
 import {resetToolBoxSizing, displayProjectName} from './editor.js';
 import {uploadHandler, uploadMergeCode} from './editor.js';
-
-import {isExperimental} from './url_parameters.js';
-import {getProjectInitialState, ProjectProfiles} from './project.js';
-import {Project, projectJsonFactory} from './project';
+import {getProjectInitialState, projectJsonFactory} from './project';
 
 // eslint-disable-next-line camelcase
 import {page_text_label} from './blockly/language/en/messages.js';
 import {utils, logConsoleMessage} from './utility';
-
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-/*  New Project Modal Dialog Handlers                                        */
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-
-/**
- * Start the process to open a new project
- *
- * @description
- * This is code that was refactored out of the editor.js
- * document.ready() handler.
- */
-export function newProjectModal() {
-  // If the current project has been modified, give the user an opportunity
-  // to abort the new project process.
-  if (isProjectChanged()) {
-    const promiseFn = () => {
-      return new Promise((resolve, reject) => {
-        utils.confirm(
-            'Abandon Current Project',
-            `The current project has been modified. Click OK to\n
-    discard the current changes and create a new project.`,
-            (result) => {
-              // result is true if the OK button was selected
-              if (result) {
-                reject( new Error('User wishes to keep existing project'));
-              } else {
-                resolve(true);
-              }
-            },
-            'Cancel',
-            'OK'
-        );
-      });
-    };
-
-    promiseFn()
-        .then(() => {
-          logConsoleMessage(`Opening New Project modal...`);
-          showNewProjectModal();
-        })
-        .catch((e) => {
-          logConsoleMessage(`ThenCatch: ${e.message}`);
-        });
-  } else {
-    // Open up a modal window to get new project details.
-    showNewProjectModal();
-  }
-}
-
-/**
- *  Displays the new project modal.
- *
- *  Modal dialog that presents user with options to provide a project
- *  name, a description and drop down menu to select a project board
- *  type.
- *
- *  The modal provides options to cancel, escape and accept the new
- *  project details. If accepted, the accept button callback will
- *  update the global project object with a new, empty project.
- */
-function showNewProjectModal() {
-  resetNewProjectModalUI();
-
-  // Set up element event handlers
-  newProjectModalCancelClick(); // Handle a click on the Cancel button
-  newProjectModalAcceptClick(); // Handle a click on the Open button
-  newProjectModalEnterClick(); // Handle the user pressing the Enter key
-  newProjectModalEscapeClick(); // Handle user clicking on the 'x' icon
-
-  populateProjectBoardTypesUIElement();
-
-  // Show the New Project modal dialog box
-  $('#new-project-dialog').modal({keyboard: false, backdrop: 'static'});
-
-  // Populate the time stamp fields
-  // TODO: These settings only apply to the Edit Project dialog
-  const projectTimestamp = new Date();
-  $('#edit-project-created-date').html(projectTimestamp);
-  $('#edit-project-last-modified').html(projectTimestamp);
-}
-
-/**
- * Reset the New Project modal input fields
- */
-function resetNewProjectModalUI() {
-  // Reset the values in the form to defaults
-  $('#new-project-name').val('');
-  $('#new-project-description').val('');
-  $('#new-project-dialog-title')
-      .html(page_text_label['editor_new_project_title']);
-}
-/**
- * Handle the <enter> keypress in the modal form
- */
-function newProjectModalEnterClick() {
-  // Ignore <enter> key pressed in a form field
-  $('#new-project-dialog').on('keydown', function(e) {
-    // Let it go if the user is in the description textarea
-    if (document.activeElement.id === 'new-project-description') {
-      return;
-    }
-
-    // The Enter key was pressed
-    if (e.key === 'Enter') {
-      if (!validateNewProjectForm()) {
-        e.preventDefault();
-        // eslint-disable-next-line no-invalid-this
-        $(this).trigger('submit');
-      } else {
-        $('#new-project-dialog').modal('hide');
-        // Clear the action cookie
-        Cookies.remove('action');
-        createNewProject();
-      }
-    }
-  });
-}
-
-/**
- *  New project modal Accept button onClick handler
- */
-function newProjectModalAcceptClick() {
-  // Click event handler. When the user clicks the 'Continue'
-  // button, validate the form
-  // --------------------------------------------------------------
-  $('#new-project-continue').on('click', function() {
-    // verify that the project contains a valid board type and project name
-    if (validateNewProjectForm()) {
-      $('#new-project-dialog').modal('hide');
-      // Clear the action cookie
-      Cookies.remove('action');
-      createNewProject();
-    }
-    // TODO: Add test for existing project before resizing
-    // use a short delay to ensure the DOM is fully ready
-    // (TODO: may not be necessary)
-    // eslint-disable-next-line no-undef
-    resetToolBoxSizing(100);
-  });
-}
-
-/**
- *  New project modal Cancel button onClick handler
- */
-function newProjectModalCancelClick() {
-  // Set up the click even handler for the "New Project" modal
-  // dialog return to the splash screen if the user clicks the
-  // cancel button
-  // ----------------------------------------------------------
-  // This is also handling the 'Edit Project Details' modal
-  // dialog box
-  // ----------------------------------------------------------
-  $('#new-project-cancel').on('click', () => {
-    // Dismiss the modal in the UX
-    $('#new-project-dialog').modal('hide');
-    // Clear the action cookie
-    Cookies.remove('action');
-
-    if (! getProjectInitialState() ||
-        typeof(getProjectInitialState().board) === 'undefined' ) {
-      // If there is no project, go to home page.
-      // window.location.href = 'index.html' + window.getAllURLParameters();
-    }
-
-    // if the project is being edited, clear the fields and close the modal
-    $('#new-project-board-dropdown').removeClass('hidden');
-    $('#edit-project-details-static').addClass('hidden');
-
-    $('#new-project-board-type').val('');
-    $('#edit-project-board-type').html('');
-    $('#edit-project-created-date').html('');
-    $('#edit-project-last-modified').html('');
-  });
-}
-
-/**
- * New project modal 'x' click handler
- *
- * This is fired when the user clicks on the 'x' to dismiss
- * the dialog.
- */
-function newProjectModalEscapeClick() {
-  // Trap the modal event that fires when the modal window is
-  // closed when the user clicks on the 'x' icon.
-  $('#new-project-dialog').on('hidden.bs.modal', function() {
-    if (!getProjectInitialState() ||
-        typeof(getProjectInitialState().board) === 'undefined') {
-      // If there is no project, go to home page.
-      // window.location.href = 'index.html';
-    }
-  });
-}
-
-/**
- * Verify that the project name and board type form fields have data
- *
- * @return {boolean} True if form contains valid data, otherwise false
- */
-function validateNewProjectForm() {
-  const projectElement = $('.proj');
-
-  // Validate the jQuery object based on these rules. Supply helpful
-  // error messages to use when a rule is violated
-  projectElement.validate({
-    rules: {
-      'new-project-name': 'required',
-      'new-project-board-type': 'required',
-    },
-    messages: {
-      'new-project-name': 'Please enter a project name',
-      'new-project-board-type': 'Please select a board type',
-    },
-  });
-
-  return !!projectElement.valid();
-}
 
 /* ------------------------------------------------------------------------- */
 /* ------------------------------------------------------------------------- */
@@ -696,70 +474,4 @@ function resetUploadImportDialog() {
   // reset the blockly toolbox sizing to ensure it renders correctly:
   // eslint-disable-next-line no-undef
   resetToolBoxSizing(100);
-}
-
-
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-/*            HELPER FUNCTIONS                                               */
-/* ------------------------------------------------------------------------- */
-/* ------------------------------------------------------------------------- */
-
-/**
- * Populate the UI Project board type drop-down list
- * @description
- * element is the <select> HTML element that will be populated with a
- * collection of possible board types
- *
- * selected is an optional string parameter
- * containing the board type in the list that should be designated as the
- * selected board type.
- */
-function populateProjectBoardTypesUIElement() {
-  const element = $('#new-project-board-type');
-  if (!element) {
-    logConsoleMessage(`Unable to find Board Type UI element.`);
-    return;
-  }
-
-  // Clear out the board type dropdown menu
-  const length = element[0].options.length;
-  for (let i = length-1; i >= 0; i--) {
-    element[0].options[i].remove();
-  }
-
-  // Populate the board type dropdown menu with a header first,
-  element.append($('<option />')
-      .val('')
-      .text(page_text_label['project_create_board_type_select'])
-      .attr('disabled', 'disabled')
-      .attr('selected', 'selected')
-  );
-
-  // then populate the dropdown with the board types
-  // defined in propc.js in the 'profile' object
-  // (except 'default', which is where the current project's type is stored)
-  for (const board in ProjectProfiles) {
-    if (Object.prototype.hasOwnProperty.call(ProjectProfiles, board)) {
-      if (board !== 'default') {
-        // Use the experimental tag to show code-only view
-        if (board !== 'propcfile' ||
-            (board === 'propcfile' &&
-                isExperimental.indexOf('propc') > -1)) {
-          if (board !== 'unknown') {
-            // Exclude the 'unknown' board type. It is used only when
-            // something has gone wrong during a project load operation
-            element.append($('<option />')
-                .val(ProjectProfiles[board].name)
-                .text(ProjectProfiles[board].description));
-          }
-        }
-      }
-
-      // Optionally set the selected option element
-      // if (selected && board === selected) {
-      //   $(element).val(selected);
-      // }
-    }
-  }
 }
